applications/rtsp-buffer/requirements.txt
Bootstrap-Flask==1.5
click==7.1.2
ffmpeg-python==0.2.0
Flask==1.1.2
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyee==8.0.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.0
